refactor: 사용하지 않는 페이지 삭제 및 카카오톡 공유 버튼 경로 수정
