+++ apps/web/src/components/midpoint-result/organisms/ParticipantBottomSheet.tsx
@@ -4,7 +4,7 @@ import { theme } from "@repo/ui/tokens";
 import * as styles from "./styles.css";
 import { AnimationBottomSheet } from "@repo/ui/components";
 import { ShareIcon } from "../atom/share/ShareIcon";
-import KakaoTalkShareButton from "@/components/share-room/molecule/KakaoTalkShareButton";
+import { KakaoTalkShareButton } from "@/components/common";
 import { KAKAO_TEMPLATE_IDS } from "@/constants/kakao-template";
 import { ReactNode } from "react";
 import { useParams, useRouter } from "next/navigation";
