--- apps/web/src/components/share-room/organisms/share-room-bottom/style.css.ts
@@ -1,10 +1,0 @@
-import { style } from "@vanilla-extract/css";
-
-export const containerStyle = style({
-  marginTop: "29px",
-});
-
-export const KakaoTalkShareButtonStyle = style({
-  width: "100%",
-  padding: "20px 16px",
-});
