--- apps/web/src/components/share-room/templates/style.css.ts
@@ -1,5 +1,0 @@
-import { style } from "@vanilla-extract/css";
-
-export const containerStyle = style({
-  height: "100dvh",
-});
