--- apps/web/src/pages/Share/Page.tsx
@@ -1,5 +1,0 @@
-import ShareRoomLayout from "@/components/share-room/templates/ShareRoomLayout";
-
-export default function Page() {
-  return <ShareRoomLayout />;
-}
